@@ -2,7 +2,7 @@ from pyodide_test_runner import run_in_pyodide
 
 
 @run_in_pyodide(packages=["Jinja2"])
-def test_jinja2():
+def test_jinja2(selenium):
     import jinja2
 
     template = jinja2.Template("Hello {{ name }}!")
