@@ -2,7 +2,7 @@ from pyodide_test_runner import run_in_pyodide
 
 
 @run_in_pyodide(packages=["cffi"])
-def test_cffi_asprintf():
+def test_cffi_asprintf(selenium):
     from cffi import FFI
 
     ffi = FFI()
