@@ -1,12 +1,6 @@
 import pytest
 from pyodide_test_runner import run_in_pyodide
 
-CHROME_FAIL_v90_MSG = (
-    "Doesn't work in chrome v89 or v90, I think because of "
-    "https://bugs.chromium.org/p/chromium/issues/detail?id=1200031. "
-    "Confirmed locally to work in v91 and v96, and to break on v90."
-)
-
 
 @pytest.mark.parametrize(
     "pattern,name,flags,expected",
@@ -19,8 +13,6 @@ CHROME_FAIL_v90_MSG = (
 )
 def test_fnmatch(selenium_module_scope, pattern, name, flags, expected):
     selenium = selenium_module_scope
-    if selenium.browser == "chrome":
-        pytest.xfail(CHROME_FAIL_v90_MSG)
     selenium.load_package("cffi_example")
     result = selenium.run(
         f"""
@@ -33,10 +25,8 @@ def test_fnmatch(selenium_module_scope, pattern, name, flags, expected):
 
 @run_in_pyodide(
     packages=["cffi_example"],
-    module_scope=True,
-    xfail_browsers={"chrome": CHROME_FAIL_v90_MSG},
 )
-def test_person():
+def test_person(selenium_module_scope):
     from cffi_example.person import Person
 
     p = Person("Alex", "Smith", 72)
