@@ -2,7 +2,7 @@ from pyodide_test_runner import run_in_pyodide
 
 
 @run_in_pyodide(packages=["cloudpickle"])
-def test_cloudpickle():
+def test_cloudpickle(selenium):
     import cloudpickle
 
     squared = lambda x: x**2
