@@ -2,7 +2,7 @@ from pyodide_test_runner import run_in_pyodide
 
 
 @run_in_pyodide(packages=["fpcast-test"])
-def test_fpcasts():
+def test_fpcasts(selenium):
     import fpcast_test
 
     fpcast_test.noargs0()
