@@ -1,8 +1,8 @@
 from pyodide_test_runner import run_in_pyodide
 
 
-@run_in_pyodide(standalone=True, packages=["jedi"])
-def test_jedi():
+@run_in_pyodide(packages=["jedi"])
+def test_jedi(selenium_standalone):
     import jedi
 
     script = jedi.Script("import json\njson.lo", path="example.py")
