@@ -8,7 +8,7 @@ from pyodide_test_runner import run_in_pyodide, spawn_web_server
 
 sys.path.append(str(Path(__file__).resolve().parent / "src"))
 
-from importlib.metadata import Distribution, PackageNotFoundError
+from importlib.metadata import PackageNotFoundError, distributions
 
 try:
     import micropip
@@ -27,8 +27,7 @@ def _mock_importlib_version(name: str) -> str:
 def _mock_importlib_distributions():
     from micropip._micropip import WHEEL_BASE
 
-    for p in WHEEL_BASE.glob("*.dist-info"):
-        yield Distribution.at(p)
+    return distributions(path=[WHEEL_BASE])
 
 
 @pytest.fixture
@@ -56,7 +55,7 @@ class Wildcard:
         return True
 
 
-def make_wheel_filename(name: str, version: str, platform: str = "generic"):
+def make_wheel_filename(name: str, version: str, platform: str = "generic") -> str:
     if platform == "generic":
         platform_str = "py3-none-any"
     elif platform == "emscripten":
@@ -73,33 +72,50 @@ class mock_fetch_cls:
     def __init__(self):
         self.releases_map = {}
         self.metadata_map = {}
+        self.top_level_map = {}
 
-    def add_pkg(
+    def add_pkg_version(
         self,
         name: str,
-        requirements: dict[str, list[str]],
+        version: str = "1.0.0",
+        *,
+        requirements: list[str] | None = None,
+        extras: dict[str, list[str]] | None = None,
         platform: str = "generic",
-    ):
-        releases = {}
-        for version, reqs in requirements.items():
-            filename = make_wheel_filename(name, version, platform)
-            releases[version] = [
-                {
-                    "filename": filename,
-                    "url": filename,
-                    "digests": {
-                        "sha256": Wildcard(),
-                    },
-                }
+        top_level: list[str] | None = None,
+    ) -> None:
+        if requirements is None:
+            requirements = []
+        if extras is None:
+            extras = {}
+        if top_level is None:
+            top_level = []
+        if name not in self.releases_map:
+            self.releases_map[name] = {"releases": {}}
+        releases = self.releases_map[name]["releases"]
+        filename = make_wheel_filename(name, version, platform)
+        releases[version] = [
+            {
+                "filename": filename,
+                "url": filename,
+                "digests": {
+                    "sha256": Wildcard(),
+                },
+            }
+        ]
+        metadata = [("Name", name), ("Version", version)] + [
+            ("Requires-Dist", req) for req in requirements
+        ]
+        for extra, reqs in extras.items():
+            metadata += [("Provides-Extra", extra)] + [
+                ("Requires-Dist", f"{req}; extra == {extra!r}") for req in reqs
             ]
-            metadata = [("Name", name), ("Version", version)] + [
-                ("Requires-Dist", req) for req in reqs
-            ]
-            self.metadata_map[filename] = metadata
-        self.releases_map[name] = {"releases": releases}
+        self.metadata_map[filename] = metadata
+        self.top_level_map[filename] = top_level
 
     async def _get_pypi_json(self, pkgname, kwargs):
         try:
+            print("_get_pypi_json", pkgname, self.releases_map[pkgname])
             return self.releases_map[pkgname]
         except KeyError as e:
             raise ValueError(
@@ -113,16 +129,23 @@ class mock_fetch_cls:
         wheel_info = WheelInfo.from_url(url)
         version = wheel_info.version
         name = wheel_info.name
-        metadata = self.metadata_map[wheel_info.filename]
+        filename = wheel_info.filename
+        metadata = self.metadata_map[filename]
         metadata_str = "\n".join(": ".join(x) for x in metadata)
+        toplevel = self.top_level_map[filename]
+        toplevel_str = "\n".join(toplevel)
+
+        metadata_dir = f"{name}-{version}.dist-info"
 
         with io.BytesIO() as tmp:
             with zipfile.ZipFile(tmp, "w", zipfile.ZIP_DEFLATED) as archive:
-                archive.writestr(f"{name}-{version}.dist-info/METADATA", metadata_str)
-                archive.writestr(
-                    f"{name}-{version}.dist-info/WHEEL",
-                    "Wheel-Version: 1.0",
-                )
+
+                def write_file(filename, contents):
+                    archive.writestr(f"{metadata_dir}/{filename}", contents)
+
+                write_file("METADATA", metadata_str)
+                write_file("WHEEL", "Wheel-Version: 1.0")
+                write_file("top_level.txt", toplevel_str)
 
             tmp.seek(0)
 
@@ -382,13 +405,17 @@ def test_install_mixed_case2(selenium_standalone_micropip, jinja2):
 
 
 @pytest.mark.asyncio
-async def test_install_keep_going(mock_fetch: mock_fetch_cls, dummy_pkg_name: str):
-    mock_fetch.add_pkg(dummy_pkg_name, {"1.0.0": ["dep1", "dep2"]})
-    mock_fetch.add_pkg("dep1", {"1.0.0": []}, "native")
-    mock_fetch.add_pkg("dep2", {"1.0.0": []}, "native")
+async def test_install_keep_going(
+    mock_fetch: mock_fetch_cls, dummy_pkg_name: str
+) -> None:
+    dep1 = f"{dummy_pkg_name}-dep1"
+    dep2 = f"{dummy_pkg_name}-dep2"
+    mock_fetch.add_pkg_version(dummy_pkg_name, requirements=[dep1, dep2])
+    mock_fetch.add_pkg_version(dep1, platform="native")
+    mock_fetch.add_pkg_version(dep2, platform="native")
 
     # report order is non-deterministic
-    msg = "(dep1|dep2).*(dep2|dep1)"
+    msg = f"({dep1}|{dep2}).*({dep2}|{dep1})"
     with pytest.raises(ValueError, match=msg):
         await micropip.install(dummy_pkg_name, keep_going=True)
 
@@ -396,11 +423,12 @@ async def test_install_keep_going(mock_fetch: mock_fetch_cls, dummy_pkg_name: st
 @pytest.mark.asyncio
 async def test_install_version_compare_prerelease(
     mock_fetch: mock_fetch_cls, dummy_pkg_name: str, mock_importlib: None
-):
+) -> None:
     version_old = "3.2.0"
     version_new = "3.2.1a1"
 
-    mock_fetch.add_pkg(dummy_pkg_name, {version_old: [], version_new: []})
+    mock_fetch.add_pkg_version(dummy_pkg_name, version_old)
+    mock_fetch.add_pkg_version(dummy_pkg_name, version_new)
 
     await micropip.install(f"{dummy_pkg_name}=={version_new}")
     await micropip.install(f"{dummy_pkg_name}>={version_old}")
@@ -413,10 +441,10 @@ async def test_install_version_compare_prerelease(
 @pytest.mark.asyncio
 async def test_install_no_deps(
     mock_fetch: mock_fetch_cls, dummy_pkg_name: str, mock_importlib: None
-):
+) -> None:
     dep_pkg_name = "dependency_dummy"
-    mock_fetch.add_pkg(dummy_pkg_name, {"1.0.0": [dep_pkg_name]})
-    mock_fetch.add_pkg(dep_pkg_name, {"1.0.0": []})
+    mock_fetch.add_pkg_version(dummy_pkg_name, requirements=[dep_pkg_name])
+    mock_fetch.add_pkg_version(dep_pkg_name)
 
     await micropip.install(dummy_pkg_name, deps=False)
 
@@ -428,13 +456,14 @@ async def test_install_no_deps(
 @pytest.mark.parametrize("pre", [True, False])
 async def test_install_pre(
     mock_fetch: mock_fetch_cls, mock_importlib: None, pre: bool, dummy_pkg_name: str
-):
+) -> None:
     version_alpha = "2.0.1a1"
     version_stable = "1.0.0"
 
     version_should_select = version_alpha if pre else version_stable
 
-    mock_fetch.add_pkg(dummy_pkg_name, {version_stable: [], version_alpha: []})
+    mock_fetch.add_pkg_version(dummy_pkg_name, version_stable)
+    mock_fetch.add_pkg_version(dummy_pkg_name, version_alpha)
     await micropip.install(dummy_pkg_name, pre=pre)
     assert micropip.list()[dummy_pkg_name].version == version_should_select
 
@@ -457,8 +486,8 @@ async def test_fetch_wheel_fail(monkeypatch):
 @pytest.mark.asyncio
 async def test_list_pypi_package(
     mock_fetch: mock_fetch_cls, mock_importlib: None, dummy_pkg_name: str
-):
-    mock_fetch.add_pkg(dummy_pkg_name, {"1.0.0": []})
+) -> None:
+    mock_fetch.add_pkg_version(dummy_pkg_name)
 
     await micropip.install(dummy_pkg_name)
     pkg_list = micropip.list()
@@ -469,8 +498,8 @@ async def test_list_pypi_package(
 @pytest.mark.asyncio
 async def test_list_wheel_package(
     mock_fetch: mock_fetch_cls, mock_importlib: None, dummy_pkg_name: str
-):
-    mock_fetch.add_pkg(dummy_pkg_name, {"1.0.0": []})
+) -> None:
+    mock_fetch.add_pkg_version(dummy_pkg_name)
     dummy_url = f"https://dummy.com/{dummy_pkg_name}-1.0.0-py3-none-any.whl"
 
     await micropip.install(dummy_url)
@@ -481,10 +510,11 @@ async def test_list_wheel_package(
 
 
 @pytest.mark.asyncio
-async def test_list_wheel_name_mismatch(mock_fetch: mock_fetch_cls, mock_importlib):
-
+async def test_list_wheel_name_mismatch(
+    mock_fetch: mock_fetch_cls, mock_importlib: None
+) -> None:
     dummy_pkg_name = "dummy-Dummy"
-    mock_fetch.add_pkg(dummy_pkg_name, {"1.0.0": []})
+    mock_fetch.add_pkg_version(dummy_pkg_name)
     dummy_url = "https://dummy.com/dummy_dummy-1.0.0-py3-none-any.whl"
 
     await micropip.install(dummy_url)
@@ -535,7 +565,7 @@ def test_list_loaded_from_js(selenium_standalone_micropip):
 
 @pytest.mark.skip_refcount_check
 @run_in_pyodide(packages=["micropip"])
-async def test_install_with_credentials():
+async def test_install_with_credentials(selenium):
     import json
     from unittest.mock import MagicMock, patch
 
@@ -561,3 +591,34 @@ async def test_install_with_credentials():
         )
 
     await call_micropip_install()
+
+
+@pytest.mark.asyncio
+async def test_freeze(
+    mock_fetch: mock_fetch_cls, dummy_pkg_name: str, mock_importlib: None
+) -> None:
+    pkg = dummy_pkg_name
+    dep1 = f"{pkg}-dep1"
+    dep2 = f"{pkg}-dep2"
+    toplevel = [["abc", "def", "geh"], ["c", "h", "i"], ["a12", "b13"]]
+
+    mock_fetch.add_pkg_version(pkg, requirements=[dep1, dep2], top_level=toplevel[0])
+    mock_fetch.add_pkg_version(dep1, top_level=toplevel[1])
+    mock_fetch.add_pkg_version(dep2, top_level=toplevel[2])
+
+    await micropip.install(pkg)
+    import json
+
+    lockfile = json.loads(micropip.freeze())
+    import pprint
+
+    pprint.pprint(lockfile["packages"])
+    pkg_metadata = lockfile["packages"][pkg]
+    dep1_metadata = lockfile["packages"][dep1]
+    dep2_metadata = lockfile["packages"][dep2]
+    assert pkg_metadata["depends"] == [dep1, dep2]
+    assert dep1_metadata["depends"] == []
+    assert dep2_metadata["depends"] == []
+    assert pkg_metadata["imports"] == toplevel[0]
+    assert dep1_metadata["imports"] == toplevel[1]
+    assert dep2_metadata["imports"] == toplevel[2]
