@@ -40,7 +40,7 @@ def test_import(selenium):
 
 
 @run_in_pyodide(packages=["opencv-python", "numpy"])
-def test_image_extensions():
+def test_image_extensions(selenium):
     import cv2 as cv
     import numpy as np
 
@@ -62,7 +62,7 @@ def test_image_extensions():
 
 
 @run_in_pyodide(packages=["opencv-python", "numpy"])
-def test_io():
+def test_io(selenium):
     import cv2 as cv
     import numpy as np
 
@@ -76,7 +76,7 @@ def test_io():
 
 
 @run_in_pyodide(packages=["opencv-python", "numpy"])
-def test_drawing():
+def test_drawing(selenium):
     import cv2 as cv
     import numpy as np
 
@@ -93,7 +93,7 @@ def test_drawing():
 
 
 @run_in_pyodide(packages=["opencv-python", "numpy"])
-def test_pixel_access():
+def test_pixel_access(selenium):
     import cv2 as cv
     import numpy as np
 
@@ -109,7 +109,7 @@ def test_pixel_access():
 
 
 @run_in_pyodide(packages=["opencv-python", "numpy"])
-def test_image_processing():
+def test_image_processing(selenium):
     import cv2 as cv
     import numpy as np
 
