@@ -2,7 +2,7 @@ from pyodide_test_runner import run_in_pyodide
 
 
 @run_in_pyodide(packages=["sharedlib-test-py"])
-def test_sharedlib():
+def test_sharedlib(selenium):
     from sharedlib_test import do_the_thing
 
     assert do_the_thing(4, 5) == 29
