@@ -4,7 +4,7 @@ from pyodide_test_runner import run_in_pyodide
 @run_in_pyodide(
     packages=["numpy", "svgwrite", "attrs", "pyrsistent", "jsonschema", "tskit"]
 )
-def test_tskit():
+def test_tskit(selenium):
     import tskit
 
     # basic test
