@@ -154,7 +154,9 @@ class JsProxy:
         """
         pass
 
-    def then(self, onfulfilled: Callable, onrejected: Callable) -> "Promise":
+    def then(
+        self, onfulfilled: Callable[[Any], Any], onrejected: Callable[[Any], Any]
+    ) -> "Promise":
         """The ``Promise.then`` API, wrapped to manage the lifetimes of the
         handlers.
 
@@ -163,7 +165,7 @@ class JsProxy:
         when the promise resolves.
         """
 
-    def catch(self, onrejected: Callable, /) -> "Promise":
+    def catch(self, onrejected: Callable[[Any], Any], /) -> "Promise":
         """The ``Promise.catch`` API, wrapped to manage the lifetimes of the
         handler.
 
@@ -172,7 +174,7 @@ class JsProxy:
         when the promise resolves.
         """
 
-    def finally_(self, onfinally: Callable, /) -> "Promise":
+    def finally_(self, onfinally: Callable[[Any], Any], /) -> "Promise":
         """The ``Promise.finally`` API, wrapped to manage the lifetimes of
         the handler.
 
@@ -189,7 +191,7 @@ class JsProxy:
 
     # Argument should be a buffer.
     # See https://github.com/python/typing/issues/593
-    def assign(self, rhs: Any, /):
+    def assign(self, rhs: Any, /) -> None:
         """Assign from a Python buffer into the JavaScript buffer.
 
         Present only if the wrapped JavaScript object is an ArrayBuffer or
@@ -198,7 +200,7 @@ class JsProxy:
 
     # Argument should be a buffer.
     # See https://github.com/python/typing/issues/593
-    def assign_to(self, to: Any, /):
+    def assign_to(self, to: Any, /) -> None:
         """Assign to a Python buffer from the JavaScript buffer.
 
         Present only if the wrapped JavaScript object is an ArrayBuffer or
@@ -221,7 +223,7 @@ class JsProxy:
         an ArrayBuffer view.
         """
 
-    def to_file(self, file: IOBase, /):
+    def to_file(self, file: IOBase, /) -> None:
         """Writes a buffer to a file.
 
         Will write the entire contents of the buffer to the current position of
@@ -245,7 +247,7 @@ class JsProxy:
         data once.
         """
 
-    def from_file(self, file: IOBase, /):
+    def from_file(self, file: IOBase, /) -> None:
         """Reads from a file into a buffer.
 
         Will try to read a chunk of data the same size as the buffer from
@@ -271,7 +273,7 @@ class JsProxy:
         data once.
         """
 
-    def _into_file(self, file: IOBase, /):
+    def _into_file(self, file: IOBase, /) -> None:
         """Will write the entire contents of a buffer into a file using
         ``canOwn : true`` without any copy. After this, the buffer cannot be
         used again.
@@ -318,7 +320,7 @@ class JsProxy:
 # from pyproxy.c
 
 
-def create_once_callable(obj: Callable, /) -> JsProxy:
+def create_once_callable(obj: Callable[..., Any], /) -> JsProxy:
     """Wrap a Python callable in a JavaScript function that can be called once.
 
     After being called the proxy will decrement the reference count
@@ -464,7 +466,7 @@ class Promise(JsProxy):
     pass
 
 
-def destroy_proxies(pyproxies: JsProxy, /):
+def destroy_proxies(pyproxies: JsProxy, /) -> None:
     """Destroy all PyProxies in a JavaScript array.
 
     pyproxies must be a JsProxy of type PyProxy[]. Intended for use with the
