@@ -228,7 +228,7 @@ def test_unpack_archive(selenium_standalone):
 
 
 @run_in_pyodide
-def test_dup_pipe():
+def test_dup_pipe(selenium):
     # See https://github.com/emscripten-core/emscripten/issues/14640
     import os
 
@@ -247,7 +247,7 @@ def test_dup_pipe():
 
 
 @run_in_pyodide
-def test_dup_temp_file():
+def test_dup_temp_file(selenium):
     # See https://github.com/emscripten-core/emscripten/issues/15012
     import os
     from tempfile import TemporaryFile
@@ -268,7 +268,7 @@ def test_dup_temp_file():
 
 
 @run_in_pyodide
-def test_dup_stdout():
+def test_dup_stdout(selenium):
     # Test redirecting stdout using low level os.dup operations.
     # This sort of redirection is used in pytest.
     import os
@@ -1149,7 +1149,7 @@ def test_sys_path0(selenium):
 
 
 @run_in_pyodide
-def test_run_js():
+def test_run_js(selenium):
     from unittest import TestCase
 
     from pyodide import run_js
