@@ -737,7 +737,7 @@ def assert_py_to_js_to_py(selenium, name):
 
 
 @run_in_pyodide
-def test_recursive_list_to_js():
+def test_recursive_list_to_js(selenium):
     x: Any = []
     x.append(x)
     from pyodide import to_js
@@ -746,7 +746,7 @@ def test_recursive_list_to_js():
 
 
 @run_in_pyodide
-def test_recursive_dict_to_js():
+def test_recursive_dict_to_js(selenium):
     x: Any = {}
     x[0] = x
     from pyodide import to_js
@@ -1426,7 +1426,7 @@ def test_buffer_format_string(selenium):
 
 
 @run_in_pyodide
-def test_object_with_null_constructor():
+def test_object_with_null_constructor(selenium):
     from unittest import TestCase
 
     from js import eval as run_js
